add TextSnapshot.ts
