+++ src/formatters/SnapshotFormatter.ts
@@ -4,7 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type {TextSnapshot, TextSnapshotNode} from '../types.js';
+import type {TextSnapshot} from '../TextSnapshot.js';
+import type {TextSnapshotNode} from '../types.js';
 
 export class SnapshotFormatter {
   #snapshot: TextSnapshot;
