+++ src/types.ts
@@ -24,17 +24,6 @@ export interface GeolocationOptions {
   longitude: number;
 }
 
-export interface TextSnapshot {
-  root: TextSnapshotNode;
-  idToNode: Map<string, TextSnapshotNode>;
-  snapshotId: string;
-  selectedElementUid?: string;
-  // It might happen that there is a selected element, but it is not part of the
-  // snapshot. This flag indicates if there is any selected element.
-  hasSelectedElement: boolean;
-  verbose: boolean;
-}
-
 export interface EmulationSettings {
   networkConditions?: string;
   cpuThrottlingRate?: number;
