+++ tests/formatters/snapshotFormatter.test.ts
@@ -10,7 +10,8 @@ import {describe, it} from 'node:test';
 import type {ElementHandle} from 'puppeteer-core';
 
 import {SnapshotFormatter} from '../../src/formatters/SnapshotFormatter.js';
-import type {TextSnapshot, TextSnapshotNode} from '../../src/types.js';
+import type {TextSnapshot} from '../../src/TextSnapshot.js';
+import type {TextSnapshotNode} from '../../src/types.js';
 
 describe('snapshotFormatter', () => {
   it('formats a snapshot with value properties', () => {
